refactor: improve model robustness and fix various schema mismatches
This commit addresses several issues across the market data and trader models:

- Fix typos: Renamed `type_filed` to `type_field`, `Mutiple` to `Multiple`, and corrected `settlement_price` and `strike_price` spellings.
- Improve Deserialization:
    - Implemented a more robust `StatusCode` deserializer to handle both numeric and string formats with better error reporting.
    - Updated `custom_date_format` to gracefully handle empty or null date strings.
    - Added `#[serde(other)]` variants to numerous enums (e.g., `AssetType`, `MarketType`, `ExchangeName`) to prevent panics on unknown API responses.
    - Applied `#[serde(default)]` to boolean fields that were previously missing default values.
- Schema Alignment:
    - Updated field types from `i64` to `f64` for price-related fields in `Underlying` and `Position`.
    - Changed `htb_quantity` from `i32` to `i64` and made several `TransactionFixedIncome` fields optional.
- Misc: Added `Display` implementations for `ServiceError` and `TransactionSubAccount` for better logging.

+++ src/model/market_data/quote_response/equity.rs
@@ -183,7 +183,7 @@ pub struct Fundamental {
     /// P/E Ratio
     pub pe_ratio: f64,
 
-    // not in schema
+    // Fields not explicitly defined in the official schema
     pub last_earnings_date: Option<chrono::DateTime<chrono::Utc>>,
 }
 
@@ -337,7 +337,7 @@ pub struct QuoteEquity {
     /// Option Risk/Volatility Measurement
     pub volatility: Option<f64>,
 
-    // not in schema
+    // Fields not explicitly defined in the official schema
     pub post_market_change: Option<f64>,
     pub post_market_percent_change: Option<f64>,
 }
@@ -373,7 +373,7 @@ pub struct ReferenceEquity {
     /// example: 100
     ///
     /// Hard to borrow quantity.
-    pub htb_quantity: Option<i32>,
+    pub htb_quantity: Option<i64>,
 
     /// example: 4.5
     ///
